// apps/world-api/src/types/event.types.ts
export enum EventOutcome {
    SUCCESS = 'success',
    FAIL = 'fail',
    BLOCKED = 'blocked',
}